requirements.txt
google-generativeai>=0.3.0
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
plotly>=5.15.0
SpeechRecognition>=3.8.0
pyttsx3>=2.90
googletrans==3.1.0a0
opencv-python>=4.8.0
mediapipe>=0.10.0
scikit-learn>=1.3.0
